requirements.txt
flask==3.0.0
flask-restx==1.3.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
marshmallow==3.21.1
python-dotenv==1.0.1